+++ flink-table/flink-sql-parser/src/main/codegen/data/Parser.tdd
@@ -97,7 +97,6 @@
     "org.apache.flink.sql.parser.ddl.SqlReset"
     "org.apache.flink.sql.parser.ddl.SqlSet"
     "org.apache.flink.sql.parser.ddl.SqlTableColumn"
-    "org.apache.flink.sql.parser.dql.SqlTableKind"
     "org.apache.flink.sql.parser.ddl.SqlTableLike"
     "org.apache.flink.sql.parser.ddl.SqlTableLike.FeatureOption"
     "org.apache.flink.sql.parser.ddl.SqlTableLike.MergingStrategy"
@@ -132,6 +131,7 @@
     "org.apache.flink.sql.parser.dql.SqlShowPartitions"
     "org.apache.flink.sql.parser.dql.SqlShowProcedures"
     "org.apache.flink.sql.parser.dql.SqlShowTables"
+    "org.apache.flink.sql.parser.dql.SqlShowTables.SqlTableKind"
     "org.apache.flink.sql.parser.dql.SqlShowColumns"
     "org.apache.flink.sql.parser.dql.SqlShowCreate"
     "org.apache.flink.sql.parser.dql.SqlShowCreateMaterializedTable"
