--- flink-table/flink-sql-parser/src/main/java/org/apache/flink/sql/parser/dql/SqlTableKind.java
@@ -1,42 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.apache.flink.sql.parser.dql;
-
-import org.apache.calcite.sql.SqlKind;
-import org.apache.calcite.sql.SqlSpecialOperator;
-
-/**
- * The kind of table. Keep in sync with {@link
- * org.apache.flink.table.catalog.CatalogBaseTable.TableKind}.
- */
-public enum SqlTableKind {
-    MATERIALIZED_TABLE(new SqlSpecialOperator("SHOW MATERIALIZED TABLES", SqlKind.OTHER)),
-    TABLE(new SqlSpecialOperator("SHOW TABLES", SqlKind.OTHER)),
-    VIEW(new SqlSpecialOperator("SHOW VIEWS", SqlKind.OTHER));
-
-    private final SqlSpecialOperator operator;
-
-    SqlTableKind(final SqlSpecialOperator operator) {
-        this.operator = operator;
-    }
-
-    public SqlSpecialOperator getOperator() {
-        return operator;
-    }
-}
